rename aside.json -> docsPath.json
